+++ alembic/versions/2023_05_09_1430-f8c3e437261b_add_materialization_job.py
@@ -9,23 +9,23 @@ Create Date: 2023-05-09 14:30:22.729559+00:00
 
 import sqlalchemy as sa
 import sqlmodel
+
 from alembic import op
 
-
 # revision identifiers, used by Alembic.
-revision = 'f8c3e437261b'
-down_revision = 'e41c021c19a6'
+revision = "f8c3e437261b"
+down_revision = "e41c021c19a6"
 branch_labels = None
 depends_on = None
 
 
 def upgrade():
     # ### commands auto generated by Alembic - please adjust! ###
-    op.add_column('materializationconfig', sa.Column('job', sa.String(), nullable=True))
+    op.add_column("materializationconfig", sa.Column("job", sa.String(), nullable=True))
     # ### end Alembic commands ###
 
 
 def downgrade():
     # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_column('materializationconfig', 'job')
+    op.drop_column("materializationconfig", "job")
     # ### end Alembic commands ###
